Move to drone, Add nginx.conf, cleanup Makefile
Currently we are building using Dockerhub, which is nice but doesn't
allow us to add testing in the future or even move the building of the
helm chart to this package. Moving to Drone will allow for that change
in the future.

Add nginx config that is compliant with running as `www-data`.