security/fundamentals: initial commit
Progress still needs to be done on the applied_cryptography page but I
think this is a good start. I would like to expand on the applied
cryptography with something that implements HMACs. My first thought for
this was a website that also had client side encryption but that seems
like a much more involved example that at some point becomes its own
unique application.